issue-14/taxi_share_list Detail Page Create
API Interface 분리

기존 ServerAPI Interface 에 retrofit 관련 함수들을 모아두고 사용했었음.

이를 각 기능별로 Interface 를 분리해 작성함.
