+++ app/src/main/java/com/example/taxishare/data/remote/apis/server/LoginAPI.kt
@@ -1,0 +1,16 @@
+/*
+ * Created by WonJongSeong on 2019-07-30
+ */
+
+package com.example.taxishare.data.remote.apis.server
+
+import com.example.taxishare.data.remote.apis.server.response.LoginRequestResponse
+import io.reactivex.Observable
+import retrofit2.http.Body
+import retrofit2.http.POST
+
+interface LoginAPI {
+
+    @POST("login")
+    fun loginRequest(@Body serverRequest: Map<String, String>): Observable<LoginRequestResponse>
+}
